ajout delete multiple sur table

+++ sources/liferay-workspace/modules/gradebook/gradebook-web/src/main/java/com/liferay/training/space/gradebook/portlet/command/DeleteAssignmentMVCActionCommand.java
@@ -25,13 +25,23 @@ public class DeleteAssignmentMVCActionCommand extends BaseMVCActionCommand {
 
 	@Override
 	protected void doProcessAction(
-		ActionRequest actionRequest, ActionResponse actionResponse)
+			ActionRequest actionRequest, ActionResponse actionResponse)
 			throws Exception {
 
-		long assignmentId = ParamUtil.getLong(actionRequest, "assignmentId");
+		String[] rowIds = ParamUtil.getParameterValues(
+				actionRequest, "rowIds"
+		);
 
-		_assignmentService.deleteAssignment(assignmentId);
-		SessionMessages.add(actionRequest, "assignment-deleted");
+		if (rowIds == null || rowIds.length == 0) {
+			return;
+		}
+
+		for (String rowId : rowIds) {
+			long assignmentId = Long.parseLong(rowId);
+			_assignmentService.deleteAssignment(assignmentId);
+		}
+
+		SessionMessages.add(actionRequest, "assignments-deleted");
 	}
 
 	@Reference(cardinality = ReferenceCardinality.MANDATORY)
